+++ src/ScottPlot/Palette.cs
@@ -14,6 +14,7 @@ namespace ScottPlot
         public static ScottPlot.Drawing.Palette Aurora => new(new ScottPlot.Drawing.Colorsets.Aurora());
         public static ScottPlot.Drawing.Palette Category10 => new(new ScottPlot.Drawing.Colorsets.Category10());
         public static ScottPlot.Drawing.Palette Category20 => new(new ScottPlot.Drawing.Colorsets.Category20());
+        public static ScottPlot.Drawing.Palette ColorblindFriendly => new(new ScottPlot.Drawing.Colorsets.ColorblindFriendly());
         public static ScottPlot.Drawing.Palette Dark => new(new ScottPlot.Drawing.Colorsets.Dark());
         public static ScottPlot.Drawing.Palette DarkPastel => new(new ScottPlot.Drawing.Colorsets.DarkPastel());
         public static ScottPlot.Drawing.Palette Frost => new(new ScottPlot.Drawing.Colorsets.Frost());
@@ -51,6 +52,7 @@ namespace ScottPlot.Drawing
         public static Palette Aurora => new(new Colorsets.Aurora());
         public static Palette Category10 => new(new Colorsets.Category10());
         public static Palette Category20 => new(new Colorsets.Category20());
+        public static Palette ColorblindFriendly => new(new Colorsets.ColorblindFriendly());
         public static Palette Dark => new(new Colorsets.Dark());
         public static Palette DarkPastel => new(new Colorsets.DarkPastel());
         public static Palette Frost => new(new Colorsets.Frost());