+++ src/cookbook/ScottPlot.Cookbook/Recipes/Palette.cs
@@ -272,4 +272,29 @@ namespace ScottPlot.Cookbook.Recipes
             plt.AxisAuto(0, 0.1);
         }
     }
+
+    public class PaletteColorblindFriendly : IRecipe
+    {
+        public string Category => "Palette";
+        public string ID => "palette_ColorblindFriendly";
+        public string Title => "Colorblind Friendly";
+        public string Description => "8-color palette that has good overall variability and " +
+            "can be differentiated by individuals with red-green color blindness. " +
+            "Colors originated from Wong 2011, https://www.nature.com/articles/nmeth.1618.pdf";
+
+        public void ExecuteRecipe(Plot plt)
+        {
+            plt.Palette = ScottPlot.Palette.ColorblindFriendly;
+
+            for (int i = 0; i < plt.Palette.Count(); i++)
+            {
+                double[] xs = DataGen.Consecutive(100);
+                double[] ys = DataGen.Sin(100, phase: -i * .5 / plt.Palette.Count());
+                plt.AddScatterLines(xs, ys, lineWidth: 3);
+            }
+
+            plt.Title($"{plt.Palette}");
+            plt.AxisAuto(0, 0.1);
+        }
+    }
 }